detection/ai_detector.py
"""
AI Violence Detection Service

This module provides functionality to detect violence in images and videos
using computer vision and machine learning techniques.
"""

import cv2
import numpy as np
from typing import Tuple, Dict
from PIL import Image
import io


class ViolenceDetector:
    """
    Violence detection using AI/ML techniques.
    
    This is a placeholder implementation that demonstrates the structure.
    In production, this would integrate with a trained ML model for violence detection.
    """
    
    def __init__(self):
        """Initialize the violence detector"""
        self.confidence_threshold = 0.5
        self.model_loaded = False
        
    def load_model(self):
        """
        Load the AI model for violence detection.
        In production, this would load a trained model (e.g., PyTorch, TensorFlow)
        """
        # Placeholder - in production, load actual model
        self.model_loaded = True
        
    def preprocess_image(self, image_data: bytes) -> np.ndarray:
        """
        Preprocess image for violence detection
        
        Args:
            image_data: Raw image bytes
            
        Returns:
            Preprocessed image array
        """
        # Convert bytes to PIL Image
        image = Image.open(io.BytesIO(image_data))
        
        # Convert to RGB if needed
        if image.mode != 'RGB':
            image = image.convert('RGB')
        
        # Convert to numpy array
        img_array = np.array(image)
        
        # Resize to standard size
        img_array = cv2.resize(img_array, (224, 224))
        
        # Normalize
        img_array = img_array.astype(np.float32) / 255.0
        
        return img_array
    
    def detect_violence_in_image(self, image_data: bytes) -> Dict[str, any]:
        """
        Detect violence in a single image
        
        Args:
            image_data: Raw image bytes
            
        Returns:
            Dictionary containing detection results
        """
        if not self.model_loaded:
            self.load_model()
        
        # Preprocess image
        processed_image = self.preprocess_image(image_data)
        
        # Placeholder detection logic
        # In production, this would use a trained model
        # For now, we'll use simple heuristics based on image properties
        
        # Calculate some basic features
        mean_intensity = np.mean(processed_image)
        std_intensity = np.std(processed_image)
        
        # Placeholder confidence calculation
        # In real implementation, this would come from the ML model
        confidence = min(0.95, max(0.05, np.random.random() * 0.3 + 0.2))
        
        # Determine severity based on confidence
        if confidence > 0.8:
            severity = 'critical'
        elif confidence > 0.6:
            severity = 'high'
        elif confidence > 0.4:
            severity = 'medium'
        else:
            severity = 'low'
        
        violence_detected = confidence > self.confidence_threshold
        
        return {
            'violence_detected': violence_detected,
            'confidence_score': float(confidence),
            'severity': severity,
            'details': {
                'mean_intensity': float(mean_intensity),
                'std_intensity': float(std_intensity),
            }
        }
    
    def detect_violence_in_video(self, video_path: str, sample_rate: int = 30) -> Dict[str, any]:
        """
        Detect violence in a video file
        
        Args:
            video_path: Path to video file
            sample_rate: Analyze every Nth frame
            
        Returns:
            Dictionary containing detection results
        """
        if not self.model_loaded:
            self.load_model()
        
        cap = cv2.VideoCapture(video_path)
        
        if not cap.isOpened():
            return {
                'error': 'Failed to open video file',
                'violence_detected': False,
                'confidence_score': 0.0
            }
        
        frame_count = 0
        analyzed_frames = 0
        max_confidence = 0.0
        violence_frames = []
        
        try:
            while cap.isOpened():
                ret, frame = cap.read()
                
                if not ret:
                    break
                
                # Sample frames
                if frame_count % sample_rate == 0:
                    # Convert frame to bytes
                    _, buffer = cv2.imencode('.jpg', frame)
                    frame_bytes = buffer.tobytes()
                    
                    # Detect violence in frame
                    result = self.detect_violence_in_image(frame_bytes)
                    
                    analyzed_frames += 1
                    
                    if result['violence_detected']:
                        violence_frames.append({
                            'frame_number': frame_count,
                            'confidence': result['confidence_score'],
                            'severity': result['severity']
                        })
                        max_confidence = max(max_confidence, result['confidence_score'])
                
                frame_count += 1
        
        finally:
            cap.release()
        
        violence_detected = len(violence_frames) > 0
        
        # Calculate overall severity
        if max_confidence > 0.8:
            severity = 'critical'
        elif max_confidence > 0.6:
            severity = 'high'
        elif max_confidence > 0.4:
            severity = 'medium'
        else:
            severity = 'low'
        
        return {
            'violence_detected': violence_detected,
            'confidence_score': float(max_confidence),
            'severity': severity,
            'total_frames': frame_count,
            'analyzed_frames': analyzed_frames,
            'violence_frames': violence_frames,
            'violence_frame_count': len(violence_frames)
        }
    
    def set_confidence_threshold(self, threshold: float):
        """
        Set the confidence threshold for violence detection
        
        Args:
            threshold: Confidence threshold (0-1)
        """
        if 0 <= threshold <= 1:
            self.confidence_threshold = threshold
        else:
            raise ValueError("Threshold must be between 0 and 1")


# Singleton instance
detector = ViolenceDetector()
