requirements.txt
Django>=6.0.1
djangorestframework>=3.16.1
Pillow>=12.1.0
opencv-python-headless>=4.12.0.88
numpy>=2.2.6
